Changed some integers to doubles and created ConfigWriter

+++ reader/src/ConfigReader.h
@@ -10,8 +10,8 @@ public:
 	void reRead();  // Reread the config file
 	void display();  // Displays current parameters
 	unsigned int getStartTime() const;  // Gets the time to start reading from the sensor
-	int getExpectedLowValue() const;  // Gets the expected lowest value the sensor should see
-	int getExpectedHighValue() const;  // Gets the expected highest value sensor should see
+	double getExpectedLowValue() const;  // Gets the expected lowest value the sensor should see
+	double getExpectedHighValue() const;  // Gets the expected highest value sensor should see
 	unsigned int getSampleRate() const;  // Gets the sample rate in Hz
 	unsigned int getTimeToRead() const;  // Gets how long the sensor should read
 
@@ -19,8 +19,8 @@ private:
 	void read();
 	const std::string configFile;
 	unsigned int startTime;
-	int expectedLowValue;
-	int expectedHighValue;
+	double expectedLowValue;
+	double expectedHighValue;
 	unsigned int sampleRate;
 	unsigned int timeToRead;
 };
